[State Sync] Create a tests module and shared_components mod for shared data structures.
