+++ state-sync/src/fuzzing.rs
@@ -5,30 +5,19 @@ use crate::{
     chunk_request::{GetChunkRequest, TargetType},
     chunk_response::{GetChunkResponse, ResponseLedgerInfo},
     coordinator::StateSyncCoordinator,
-    executor_proxy::{ExecutorProxy, ExecutorProxyTrait},
-    network::{StateSyncMessage, StateSyncSender},
+    executor_proxy::ExecutorProxy,
+    network::StateSyncMessage,
+    shared_components::test_utils,
 };
-use channel::{diem_channel, message_queues::QueueStyle};
 use diem_config::{
-    config::{PeerNetworkId, RoleType, StateSyncConfig, UpstreamConfig},
+    config::PeerNetworkId,
     network_id::{NetworkId, NodeNetworkId},
 };
 use diem_infallible::Mutex;
 use diem_types::{
-    ledger_info::LedgerInfoWithSignatures,
-    transaction::{Transaction, TransactionListWithProof, WriteSetPayload},
-    waypoint::Waypoint,
-    PeerId,
+    ledger_info::LedgerInfoWithSignatures, transaction::TransactionListWithProof, PeerId,
 };
-use diem_vm::DiemVM;
-use diemdb::DiemDB;
-use executor::Executor;
-use executor_test_helpers::bootstrap_genesis;
-use futures::{channel::mpsc, executor::block_on};
-use network::{
-    peer_manager::{ConnectionRequestSender, PeerManagerRequestSender},
-    protocols::network::NewNetworkSender,
-};
+use futures::executor::block_on;
 use once_cell::sync::Lazy;
 use proptest::{
     arbitrary::{any, Arbitrary},
@@ -36,61 +25,9 @@ use proptest::{
     prelude::*,
     strategy::Strategy,
 };
-use std::collections::HashMap;
-use storage_interface::DbReaderWriter;
 
-static STATE_SYNC_COORDINATOR: Lazy<Mutex<StateSyncCoordinator<ExecutorProxy>>> = Lazy::new(|| {
-    // Generate a genesis change set
-    let (genesis, _) = vm_genesis::test_genesis_change_set_and_validators(Some(1));
-
-    // Create test diem database
-    let db_path = diem_temppath::TempPath::new();
-    db_path.create_as_dir().unwrap();
-    let (db, db_rw) = DbReaderWriter::wrap(DiemDB::new_for_test(db_path.path()));
-
-    // Bootstrap the genesis transaction
-    let genesis_txn = Transaction::GenesisTransaction(WriteSetPayload::Direct(genesis));
-    bootstrap_genesis::<DiemVM>(&db_rw, &genesis_txn).unwrap();
-
-    // Create executor proxy
-    let chunk_executor = Box::new(Executor::<DiemVM>::new(db_rw));
-    let executor_proxy = ExecutorProxy::new(db, chunk_executor, vec![]);
-
-    // Get initial state
-    let initial_state = executor_proxy.get_local_storage_state().unwrap();
-
-    // Setup network senders
-    let (network_reqs_tx, _network_reqs_rx) = diem_channel::new(QueueStyle::FIFO, 8, None);
-    let (connection_reqs_tx, _) = diem_channel::new(QueueStyle::FIFO, 8, None);
-    let network_sender = StateSyncSender::new(
-        PeerManagerRequestSender::new(network_reqs_tx),
-        ConnectionRequestSender::new(connection_reqs_tx),
-    );
-    let node_network_id = NodeNetworkId::new(NetworkId::Validator, 0);
-    let network_senders = vec![(node_network_id, network_sender)]
-        .into_iter()
-        .collect::<HashMap<_, _>>();
-
-    // Create channel senders and receivers
-    let (_coordinator_sender, coordinator_receiver) = mpsc::unbounded();
-    let (mempool_sender, _mempool_receiver) = mpsc::channel(1);
-
-    // Start up coordinator
-    let coordinator = StateSyncCoordinator::new(
-        coordinator_receiver,
-        mempool_sender,
-        network_senders,
-        RoleType::Validator,
-        Waypoint::default(),
-        StateSyncConfig::default(),
-        UpstreamConfig::default(),
-        executor_proxy,
-        initial_state,
-    )
-    .unwrap();
-
-    Mutex::new(coordinator)
-});
+static STATE_SYNC_COORDINATOR: Lazy<Mutex<StateSyncCoordinator<ExecutorProxy>>> =
+    Lazy::new(|| Mutex::new(test_utils::create_state_sync_coordinator_for_tests()));
 
 proptest! {
     #![proptest_config(ProptestConfig::with_cases(10))]
