+++ state-sync/src/lib.rs
@@ -17,6 +17,7 @@ pub mod executor_proxy;
 mod logging;
 pub mod network;
 mod request_manager;
+pub mod shared_components;
 
 #[cfg(any(feature = "fuzzing", test))]
 pub mod fuzzing;
